Refactor: Split houdini/traverser.py to extract output detection logic
- Created houdini/output_detector.py containing all renderer-specific output detection functions
- Modified houdini/traverser.py to remove the five _detect_* static methods
- Updated create_output_dict to delegate to output_detector.detect_output_nodes()
- Maintains full backward compatibility through thin wrapper
- Addresses monolith module issue in houdini/traverser.py (reduced from 906 to 638 lines)
